Refactors model form types
Splits the combined type definition file into separate files
for better organization and maintainability.
