+++ plugins/main/public/components/overview/dashboard-assistant/components/types/index.ts
@@ -1,0 +1,2 @@
+export * from './model-form-data';
+export * from './model-field-definition';
