+++ plugins/main/public/components/overview/dashboard-assistant/components/types/model-field-definition.ts
@@ -1,10 +1,3 @@
-export interface ModelFormData {
-  modelProvider: string;
-  model: string;
-  apiUrl: string;
-  apiKey: string;
-}
-
 export interface ModelFieldDefinition {
   name: string;
   id: string;
