+++ plugins/main/public/components/overview/dashboard-assistant/components/types/model-form-data.ts
@@ -1,0 +1,6 @@
+export interface ModelFormData {
+  modelProvider: string;
+  model: string;
+  apiUrl: string;
+  apiKey: string;
+}
